+++ CHANGELOG.md
@@ -1,15 +1,20 @@
 # Change Log
 
+ping route
+
 ## Unreleased
+ - 
 
 ### Added
  - Load datasource route
  - Unload datasource route
  - View loaded datasources route
  - Gracefull shutdown. Waits to shut down app until all websockets are disconnected
+ - server profile route (uptime, runtime, server status)
 
 ### Changed
  - get_requirements.sh checks for src packages before downloading
+ - json response from delete layer fix
 
 
 ## 1.4.0 - 2016-03-09